feat: create component search and states
