+++ src/components/ProgressBar/styles.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const BarStatus = styled.progress`
+  height: 10px;
+  border-radius: 10px;
+
+  ::-webkit-progress-value {
+    background-color: red;
+  }
+`;
