+++ styles/globals.css
@@ -8,5 +8,5 @@ html,
 body {
   max-width: 100vw;
   overflow-x: hidden;
-  background: black;
+  background: #191621;
 }
